requirements.txt
accelerate==0.16.0
black==23.1.0
einops==0.6.0
ema-pytorch==0.1.4
isort==5.12.0
matplotlib==3.7.0
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
PyYAML==6.0
seaborn==0.12.2
torch==1.13.1
torchinfo==1.7.2
torchvision==0.14.1
tqdm==4.64.1
